Replace linear search with grade and binsearch
Switch out the searching implementation with one precomputing the
Eytzinger layout for the sorted `x` argument, while saving the
appropriate permutation to restore the correct indices.

The Eytzinger layout generation and searching code is mostly a direct translation
from [this Algorithmica article](https://algorithmica.org/en/eytzinger).

This should introduce a small regression in performance especially for very short `y` but
is asymptotically faster on large `x` and `y`.