compute last uses on the fly
By default, when computing use of a local variable, assume that it is a
last use and emit a `bd` instruction.
To support this, keep a list `lu` of indices in the bytecode array `b`.
At any point during compilation, `lu` holds the positions of all emitted
`bd` opcodes in an execution path that may flow into the one we are
currently compiling, without passing through an intermediate assignment.

- when assigning a new value to a local, remove from `lu` the index of any
 instruction that is a use of that local. (by `rl(i,0)`)
- when getting a local for any use (modified assignment or pushing it on
 the stack), iterate through the list patching the `bd`s to `bg`s so that
 the value held by the local survives to this point in the execution path.
 (by `rl(i,1)`)
- if the use is for pushing the local on the stack (thus emitting `bd`),
 also add the offset of the instruction being emitted onto `lu`.
 (by `rl(i,2)`)
- when branching, back up `lu`.
- when reconciliating two branches, take the `lu` from each and compute
  their union as the new `lu`.
- when returning early due to `:`, empty `lu`.

further additions:
- propagate the `r` flag to the branches of `:[i;t;e]`
- fix incorrect code generation due to some uses of blocks
  having a dirty source offset